+++ src/main/java/com/nanonaitor/arsenal/NanonaitorsArsenal.java
@@ -2,6 +2,7 @@ package com.nanonaitor.arsenal;
 
 import com.nanonaitor.arsenal.client.ArsenalCreativeTab;
 import com.nanonaitor.arsenal.compat.SilverSetBonusCompat;
+import com.nanonaitor.arsenal.compat.ReskillableCompat;
 import com.nanonaitor.arsenal.network.ModNetwork;
 import com.nanonaitor.arsenal.recipe.ModSmeltingRecipes;
 import net.minecraft.creativetab.CreativeTabs;
@@ -11,6 +12,7 @@ import org.apache.logging.log4j.Logger;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
+import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLServerStartedEvent;
 
 @Mod(
@@ -18,12 +20,12 @@ import net.minecraftforge.fml.common.event.FMLServerStartedEvent;
     name = NanonaitorsArsenal.NAME,
     version = NanonaitorsArsenal.VERSION,
     acceptedMinecraftVersions = "[1.12.2]",
-    dependencies = "after:setbonus;after:xat;after:quark"
+    dependencies = "after:setbonus;after:xat;after:quark;after:spartanweaponry;after:spartanfire;after:srparasites;after:reskillable"
 )
 public final class NanonaitorsArsenal {
     public static final String MOD_ID = "nanonaitors_arsenal";
     public static final String NAME = "Nanonaitor's Arsenal";
-    public static final String VERSION = "1.1.5";
+    public static final String VERSION = "1.1.6";
     public static final Logger LOGGER = LogManager.getLogger(NAME);
     public static final CreativeTabs CREATIVE_TAB = new ArsenalCreativeTab();
 
@@ -36,6 +38,11 @@ public final class NanonaitorsArsenal {
     @Mod.EventHandler
     public void init(FMLInitializationEvent event) {
         ModSmeltingRecipes.register();
+    }
+
+    @Mod.EventHandler
+    public void postInit(FMLPostInitializationEvent event) {
+        ReskillableCompat.registerWeaponRequirements();
     }
 
     @Mod.EventHandler